+++ kernel/include/algo/dynamic_array.h
@@ -15,23 +15,23 @@
 struct dynamic_array_bucket {
     void* data;
     struct dynamic_array_bucket* next;
-    uint32_t capacity;
-    uint32_t size;
+    size_t capacity;
+    size_t size;
 };
 typedef struct dynamic_array_bucket dynamic_array_bucket_t;
 
 struct dynamic_array {
     dynamic_array_bucket_t* head;
     dynamic_array_bucket_t* tail;
-    uint32_t size; /* number of elements in vector */
-    uint32_t element_size; /* size of elements in bytes */
+    size_t size; /* number of elements in vector */
+    size_t element_size; /* size of elements in bytes */
 };
 typedef struct dynamic_array dynamic_array_t;
 
 #define dynarr_init(type, v) dynarr_init_of_size_impl(v, sizeof(type), 8)
 #define dynarr_init_of_size(type, v, cap) dynarr_init_of_size_impl(v, sizeof(type), cap)
 
-int dynarr_init_of_size_impl(dynamic_array_t* v, uint32_t element_size, uint32_t capacity);
+int dynarr_init_of_size_impl(dynamic_array_t* v, size_t element_size, size_t capacity);
 int dynarr_free(dynamic_array_t* v);
 
 void* dynarr_get(dynamic_array_t* v, int index);
