+++ kernel/include/mem/kmalloc.h
@@ -16,11 +16,11 @@
 #define KMALLOC_BLOCK_SIZE 32
 
 void kmalloc_init();
-void* kmalloc(uint32_t size);
-void* kmalloc_aligned(uint32_t size, uint32_t alignment);
+void* kmalloc(size_t size);
+void* kmalloc_aligned(size_t size, size_t alignment);
 void* kmalloc_page_aligned();
 void kfree(void* ptr);
 void kfree_aligned(void* ptr);
-void* krealloc(void* ptr, uint32_t size);
+void* krealloc(void* ptr, size_t size);
 
 #endif // _KERNEL_MEM_KMALLOC_H
