+++ kernel/include/tasking/elf.h
@@ -177,6 +177,6 @@ struct file_descriptor;
 int elf_check_header(elf_header_32_t* header);
 int elf_load(struct proc* p, struct file_descriptor* fd);
 int elf_find_symtab_unchecked(void* mapped_data, void** symtab, size_t* symtab_entries, char** strtab);
-ssize_t elf_find_function_in_symtab(void* symtab, size_t syms_n, uint32_t ip);
+ssize_t elf_find_function_in_symtab(void* symtab, size_t syms_n, uintptr_t ip);
 
 #endif // _KERNEL_TASKING_ELF_H
