+++ kernel/kernel/tasking/proc_zoner.c
@@ -16,18 +16,18 @@
  * PROC ZONING
  */
 
-static inline bool _proc_zones_intersect(uint32_t start1, uint32_t size1, uint32_t start2, uint32_t size2)
+static inline bool _proc_zones_intersect(size_t start1, size_t size1, size_t start2, size_t size2)
 {
-    uint32_t end1 = start1 + size1 - 1;
-    uint32_t end2 = start2 + size2 - 1;
+    size_t end1 = start1 + size1 - 1;
+    size_t end2 = start2 + size2 - 1;
     return (start1 <= start2 && start2 <= end1) || (start1 <= end2 && end2 <= end1) || (start2 <= start1 && start1 <= end2) || (start2 <= end1 && end1 <= end2);
 }
 
-static inline bool _proc_can_fixup_zone(proc_t* proc, uint32_t* start_ptr, int* len_ptr)
+static inline bool _proc_can_fixup_zone(proc_t* proc, size_t* start_ptr, int* len_ptr)
 {
-    uint32_t zones_count = proc->zones.size;
+    size_t zones_count = proc->zones.size;
 
-    for (uint32_t i = 0; i < zones_count; i++) {
+    for (size_t i = 0; i < zones_count; i++) {
         proc_zone_t* zone = (proc_zone_t*)dynarr_get(&proc->zones, i);
         if (_proc_zones_intersect(*start_ptr, *len_ptr, zone->start, zone->len)) {
             if (*start_ptr >= zone->start) {
@@ -48,11 +48,11 @@ static inline bool _proc_can_fixup_zone(proc_t* proc, uint32_t* start_ptr, int* 
     return true;
 }
 
-static inline bool _proc_can_add_zone(proc_t* proc, uint32_t start, uint32_t len)
+static inline bool _proc_can_add_zone(proc_t* proc, size_t start, size_t len)
 {
-    uint32_t zones_count = proc->zones.size;
+    size_t zones_count = proc->zones.size;
 
-    for (uint32_t i = 0; i < zones_count; i++) {
+    for (size_t i = 0; i < zones_count; i++) {
         proc_zone_t* zone = (proc_zone_t*)dynarr_get(&proc->zones, i);
         if (_proc_zones_intersect(start, len, zone->start, zone->len)) {
             return false;
@@ -83,7 +83,7 @@ static inline void _proc_swap_zones(proc_zone_t* one, proc_zone_t* two)
 /**
  * Inserts zone, which won't overlap with existing ones.
  */
-proc_zone_t* proc_extend_zone(proc_t* proc, uint32_t start, uint32_t len)
+proc_zone_t* proc_extend_zone(proc_t* proc, size_t start, size_t len)
 {
     len += (start & (VMM_PAGE_SIZE - 1));
     start &= ~(VMM_PAGE_SIZE - 1);
@@ -107,7 +107,7 @@ proc_zone_t* proc_extend_zone(proc_t* proc, uint32_t start, uint32_t len)
     return NULL;
 }
 
-proc_zone_t* proc_new_zone(proc_t* proc, uint32_t start, uint32_t len)
+proc_zone_t* proc_new_zone(proc_t* proc, size_t start, size_t len)
 {
     len += (start & (VMM_PAGE_SIZE - 1));
     start &= ~(VMM_PAGE_SIZE - 1);
@@ -131,13 +131,13 @@ proc_zone_t* proc_new_zone(proc_t* proc, uint32_t start, uint32_t len)
 }
 
 /* FIXME: Think of more efficient way */
-proc_zone_t* proc_new_random_zone(proc_t* proc, uint32_t len)
+proc_zone_t* proc_new_random_zone(proc_t* proc, size_t len)
 {
     if (len % VMM_PAGE_SIZE) {
         len += VMM_PAGE_SIZE - (len % VMM_PAGE_SIZE);
     }
 
-    uint32_t zones_count = proc->zones.size;
+    size_t zones_count = proc->zones.size;
 
     /* Check if we can put it at the beginning */
     proc_zone_t* ret = proc_new_zone(proc, 0, len);
@@ -145,9 +145,9 @@ proc_zone_t* proc_new_random_zone(proc_t* proc, uint32_t len)
         return ret;
     }
 
-    uint32_t min_start = 0xffffffff;
+    size_t min_start = 0xffffffff;
 
-    for (uint32_t i = 0; i < zones_count; i++) {
+    for (size_t i = 0; i < zones_count; i++) {
         proc_zone_t* zone = (proc_zone_t*)dynarr_get(&proc->zones, i);
         if (_proc_can_add_zone(proc, zone->start + zone->len, len)) {
             if (min_start > zone->start + zone->len) {
@@ -164,13 +164,13 @@ proc_zone_t* proc_new_random_zone(proc_t* proc, uint32_t len)
 }
 
 /* FIXME: Think of more efficient way */
-proc_zone_t* proc_new_random_zone_backward(proc_t* proc, uint32_t len)
+proc_zone_t* proc_new_random_zone_backward(proc_t* proc, size_t len)
 {
     if (len % VMM_PAGE_SIZE) {
         len += VMM_PAGE_SIZE - (len % VMM_PAGE_SIZE);
     }
 
-    uint32_t zones_count = proc->zones.size;
+    size_t zones_count = proc->zones.size;
 
     /* Check if we can put it at the end */
     proc_zone_t* ret = proc_new_zone(proc, KERNEL_BASE - len, len);
@@ -178,9 +178,9 @@ proc_zone_t* proc_new_random_zone_backward(proc_t* proc, uint32_t len)
         return ret;
     }
 
-    uint32_t max_end = 0;
+    size_t max_end = 0;
 
-    for (uint32_t i = 0; i < zones_count; i++) {
+    for (size_t i = 0; i < zones_count; i++) {
         proc_zone_t* zone = (proc_zone_t*)dynarr_get(&proc->zones, i);
         if (_proc_can_add_zone(proc, zone->start - len, len)) {
             if (max_end < zone->start) {
@@ -196,11 +196,11 @@ proc_zone_t* proc_new_random_zone_backward(proc_t* proc, uint32_t len)
     return proc_new_zone(proc, max_end - len, len);
 }
 
-proc_zone_t* proc_find_zone_no_proc(dynamic_array_t* zones, uint32_t addr)
+proc_zone_t* proc_find_zone_no_proc(dynamic_array_t* zones, size_t addr)
 {
-    uint32_t zones_count = zones->size;
+    size_t zones_count = zones->size;
 
-    for (uint32_t i = 0; i < zones_count; i++) {
+    for (size_t i = 0; i < zones_count; i++) {
         proc_zone_t* zone = (proc_zone_t*)dynarr_get(zones, i);
         if (zone->start <= addr && addr < zone->start + zone->len) {
             return zone;
@@ -210,16 +210,16 @@ proc_zone_t* proc_find_zone_no_proc(dynamic_array_t* zones, uint32_t addr)
     return NULL;
 }
 
-proc_zone_t* proc_find_zone(proc_t* proc, uint32_t addr)
+proc_zone_t* proc_find_zone(proc_t* proc, size_t addr)
 {
     return proc_find_zone_no_proc(&proc->zones, addr);
 }
 
 int proc_delete_zone_no_proc(dynamic_array_t* zones, proc_zone_t* givzone)
 {
-    uint32_t zones_count = zones->size;
+    size_t zones_count = zones->size;
 
-    for (uint32_t i = 0; i < zones_count; i++) {
+    for (size_t i = 0; i < zones_count; i++) {
         proc_zone_t* zone = (proc_zone_t*)dynarr_get(zones, i);
         if (givzone == zone) {
             _proc_swap_zones(zone, dynarr_get(zones, zones_count - 1));
